Fixing up unit tests (#72)
The only unit test that existed did not work, so I fixed that and added
tests for most other resources in this module. I did not add tests for
the anchors as the anchor pattern is not considered good practice and is
something that `pdk validate` complains about.

I also added an integration acceptance test. This test actually applies
the graylog::server class per the documentation to test that the module
behaves as expected. This test could be expanded upon, but this serves
as a good starting point.